feat(web): stacking dao self-claim disclosure
